+++ src/graph.rs
@@ -1,4 +1,5 @@
 use crate::basic_block::*;
+use crate::batched_basic_block::*;
 use ark_bn254::{Fr, G1Affine, G1Projective, G2Affine, G2Projective};
 use ndarray::ArrayD;
 use rand::rngs::StdRng;
